refacto: initialization NF structure
